* fixed a bunch of signed/unsigned comparison warnings
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@13766 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ graphics/BatchedMeshTests.cpp
@@ -141,7 +141,7 @@ static void perfTestWithMesh(const std::string& path)
 
 	const int compression_levels[] ={ 1, 3, 6 };// , 9, 20};
 
-	for(int i=0; i<staticArrayNumElems(compression_levels); ++i)
+	for(size_t i=0; i<staticArrayNumElems(compression_levels); ++i)
 	{
 		write_options.compression_level = compression_levels[i];
